Export Data to CSV and Email vis SMTP
Modified the program to be able to export the game data as CSV in the format:
[Game ID, Round Number, Trade Price, Buyer Card, Seller Card, BuyerID, SellerID, Tax Value, Ceiling Value, Floor Value]
The host can then choose to email the data by entering an email at any stage of the game before ending (roundSettings.js)
Right now the CSV is only created if the host wishes to export the data - in it's own function though so can be separated if required.
Previous games CSVs are removed if a new game is started.

Added new module 'send_email' to contain all the SMTP info.
Will need to create a config folder to be kept locally (not uploaded to github) containing the password for the email account 'Pitmarketgame@gmail.com'