Bug fixes: Scrolling offers
Added in overflow for the <div> containter for the offers/transactions.

Offers and transactions should now have a scroll bar so that the offers don't continuously go down the page.

Also edited main.js to change colour of the button when a trade is confirmed to make it more apparent.